# Update Definition Helper Refactor