--- a/src/MongoFramework/Infrastructure/Commands/UpdateEntityCommand.cs
+++ b/src/MongoFramework/Infrastructure/Commands/UpdateEntityCommand.cs
@@ -1,7 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using MongoDB.Driver;
+using MongoFramework.Bson;
 using MongoFramework.Infrastructure.DefinitionHelpers;
 using MongoFramework.Infrastructure.Mapping;
 
@@ -18,13 +17,12 @@ namespace MongoFramework.Infrastructure.Commands
 
 		public IEnumerable<WriteModel<TEntity>> GetModel()
 		{
-			var definition = EntityMapping.GetOrCreateDefinition(typeof(TEntity));
-			var updateDefintion = UpdateDefinitionHelper.CreateFromDiff<TEntity>(EntityEntry.OriginalValues, EntityEntry.CurrentValues);
-
 			//MongoDB doesn't like it if an UpdateDefinition is empty.
 			//This is primarily to work around a mutation that may set an entity to its default state.
-			if (updateDefintion.HasChanges())
+			if (BsonDiff.HasDifferences(EntityEntry.OriginalValues, EntityEntry.CurrentValues))
 			{
+				var definition = EntityMapping.GetOrCreateDefinition(typeof(TEntity));
+				var updateDefintion = UpdateDefinitionHelper.CreateFromDiff<TEntity>(EntityEntry.OriginalValues, EntityEntry.CurrentValues);
 				yield return new UpdateOneModel<TEntity>(definition.CreateIdFilterFromEntity(EntityEntry.Entity), updateDefintion);
 			}
 		}