This reduces allocations while also avoiding reflection needed to create update definitions.

Before
|    Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| Benchmark | 67.31 us | 1.135 us | 1.006 us | 6.7139 |     - |     - |  20.64 KB |

After
|    Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| Benchmark | 22.56 us | 0.274 us | 0.229 us | 3.0518 |     - |     - |   9.43 KB |